requirements.txt
Flask
upstox-python-sdk
python-dotenv
gunicorn